requirements.txt
# =============================================================================
# Neoplast Lead Dashboard — Production dependencies
# Pinned to exact versions for reproducible builds.
# =============================================================================

# --- Web framework ----------------------------------------------------------
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# --- Validation / config ----------------------------------------------------
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# --- Database ---------------------------------------------------------------
sqlalchemy==2.0.36
alembic==1.14.0

# --- Auth / security --------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# --- Data processing --------------------------------------------------------
pandas==2.2.3
openpyxl==3.1.5

# --- HTTP / external APIs ---------------------------------------------------
httpx==0.28.1
tenacity==9.0.0

# --- Logging / utility ------------------------------------------------------
python-dotenv==1.0.1
structlog==24.4.0
